moved measles test from structureMapUtilitiesTest to FhirCarePlanGeneratorTest
